Appearance Updates
Basic positioning is being worked out. Need to get a bunch of
screenshots to use as mock ups. It will be easier to avoid making things
scale so things will be fixed at a 16:9 aspect ratio with a 1024x576
resolution. The rest of the space will be used for more in depth
character information.